+++ cli/integration-test/sample-test/src/sample_test/implementation/a.clj
@@ -1,0 +1,4 @@
+(ns sample-test.implementation.a)
+
+(defprotocol Foo
+  (something [_this]))